# Implement branchless inode::prepend_key_prefix