Baseline performance:

2020-06-26 06:25:10
Running ./micro_benchmark_key_prefix
Run on (8 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x4)
  L1 Instruction 32 KiB (x4)
  L2 Unified 256 KiB (x4)
  L3 Unified 8192 KiB (x1)
Load Average: 0.77, 0.42, 0.16
-------------------------------------------------------------------------------------------
Benchmark                                 Time             CPU   Iterations UserCounters...
-------------------------------------------------------------------------------------------
unpredictable_prepend_key_prefix       23.7 us         23.6 us        29611 items_per_second=10.6629M/s

With the change:

unpredictable_prepend_key_prefix       19.2 us         19.2 us        36497 items_per_second=13.1389M/s

perf stat baseline:

          4,530.88 msec task-clock                #    0.996 CPUs utilized
                 6      context-switches          #    0.001 K/sec
                 0      cpu-migrations            #    0.000 K/sec
               186      page-faults               #    0.041 K/sec
    15,430,018,979      cycles                    #    3.406 GHz                      (83.36%)
     3,401,693,096      stalled-cycles-frontend   #   22.05% frontend cycles idle     (83.31%)
     1,548,841,285      stalled-cycles-backend    #   10.04% backend cycles idle      (66.63%)
    33,795,109,289      instructions              #    2.19  insn per cycle
                                                  #    0.10  stalled cycles per insn  (83.31%)
     6,940,190,498      branches                  # 1531.752 M/sec                    (83.31%)
        51,949,974      branch-misses             #    0.75% of all branches          (83.38%)

       4.549329315 seconds time elapsed

       4.479173000 seconds user
       0.052036000 seconds sys

With the change:

          4,947.91 msec task-clock                #    0.996 CPUs utilized
                 4      context-switches          #    0.001 K/sec
                 0      cpu-migrations            #    0.000 K/sec
               185      page-faults               #    0.037 K/sec
    16,850,794,694      cycles                    #    3.406 GHz                      (83.31%)
     3,399,379,919      stalled-cycles-frontend   #   20.17% frontend cycles idle     (83.31%)
     1,322,476,773      stalled-cycles-backend    #    7.85% backend cycles idle      (66.69%)
    39,091,678,006      instructions              #    2.32  insn per cycle
                                                  #    0.09  stalled cycles per insn  (83.35%)
     7,967,971,347      branches                  # 1610.371 M/sec                    (83.35%)
        30,196,745      branch-misses             #    0.38% of all branches          (83.34%)

       4.966332324 seconds time elapsed

       4.884147000 seconds user
       0.064054000 seconds sys